+++ src/Node.ts
@@ -6,6 +6,7 @@ import { UpdateCommonProperties } from "./util/UpdateCommonProps";
 export interface IClassNode {
   remove(): Promise<IHttpStatus>;
   update(arg: INodeUpdate): Promise<IHttpStatus>;
+  setCentral(): Promise<IHttpStatus>;
   details: IServerNodeConfiguration;
 }
 
@@ -74,6 +75,12 @@ export class Node implements IClassNode {
 
     return await this.repoClient
       .Put(`servernodeconfiguration/${this.details.id}`, { ...this.details })
+      .then((res) => res.status);
+  }
+
+  public async setCentral() {
+    return await this.repoClient
+      .Get(`failover/tonode/${this.details.id}`)
       .then((res) => res.status);
   }
 }