+++ src/ReloadTaskBase.ts
@@ -12,6 +12,7 @@ import { UpdateCommonProperties } from "./util/UpdateCommonProps";
 import { IClassSchemaTrigger, SchemaTrigger } from "./SchemaTrigger";
 import { CompositeTrigger, IClassCompositeTrigger } from "./CompositeTrigger";
 import { TDaysOfMonth, TDaysOfWeek, TRepeatOptions } from "./types/ranges";
+import { schemaRepeat } from "./util/schemaTrigger";
 
 export interface IClassReloadTask {
   remove(): Promise<IHttpStatus>;
@@ -262,7 +263,7 @@ export abstract class ReloadTaskBase implements IClassReloadTask {
 
     let currentTimeStamp = new Date();
 
-    let schemaRepeatOpt = this.schemaRepeat(
+    let schemaRepeatOpt = schemaRepeat(
       arg.repeat || "Daily",
       arg.repeatEvery || 1,
       arg.daysOfWeek || ["Monday"],
@@ -378,63 +379,6 @@ export abstract class ReloadTaskBase implements IClassReloadTask {
       });
 
     return selectionData;
-  }
-
-  private schemaRepeat(
-    repeat: TRepeatOptions,
-    repeatEvery: number,
-    daysOfWeek: TDaysOfWeek[],
-    daysOfMonth: TDaysOfMonth[]
-  ): { incrementDescr: string; schemaFilterDescr: string } {
-    if (repeat == "Once")
-      return {
-        incrementDescr: "0 0 0 0",
-        schemaFilterDescr: "* * - * * * * *",
-      };
-
-    if (repeat == "Minute")
-      return {
-        incrementDescr: `${repeatEvery} 0 0 0`,
-        schemaFilterDescr: "* * - * * * * *",
-      };
-
-    if (repeat == "Hourly")
-      return {
-        incrementDescr: `0 ${repeatEvery} 0 0`,
-        schemaFilterDescr: "* * - * * * * *",
-      };
-
-    if (repeat == "Daily")
-      return {
-        incrementDescr: `0 0 ${repeatEvery} 0`,
-        schemaFilterDescr: "* * - * * * * *",
-      };
-
-    if (repeat == "Weekly") {
-      let weekDay = this.getWeekDayNumber(daysOfWeek);
-      return {
-        incrementDescr: `0 0 1 0`,
-        schemaFilterDescr: `* * - ${weekDay} ${repeatEvery} * * *`,
-      };
-    }
-
-    if (repeat == "Monthly")
-      return {
-        incrementDescr: `0 0 1 0`,
-        schemaFilterDescr: `* * - * * ${daysOfMonth} * *`,
-      };
-  }
-
-  private getWeekDayNumber(daysOfWeek: TDaysOfWeek[]): number[] {
-    return daysOfWeek.map((d) => {
-      if (d == "Sunday") return 0;
-      if (d == "Monday") return 1;
-      if (d == "Tuesday") return 2;
-      if (d == "Wednesday") return 3;
-      if (d == "Thursday") return 4;
-      if (d == "Friday") return 5;
-      if (d == "Saturday") return 6;
-    });
   }
 
   private async getTaskDetails() {